src/reliability.py
```python
import logging
from src.utils import send_error_report

def ensure_reliability():
    try:
        # Call all the functions of the AI agent
        curate_content()
        optimize_seo()
        interact_with_user()
        automate_publishing()
        edit_and_review()
    except Exception as e:
        logging.error(f"An error occurred: {str(e)}")
        send_error_report(str(e))
        return False
    return True

def curate_content():
    # This function will be implemented in src/content_curation.py
    pass

def optimize_seo():
    # This function will be implemented in src/seo_optimization.py
    pass

def interact_with_user():
    # This function will be implemented in src/user_interaction.py
    pass

def automate_publishing():
    # This function will be implemented in src/automation.py
    pass

def edit_and_review():
    # This function will be implemented in src/editing_reviewing.py
    pass
```